feat: onback added to itemdetilpage
