+++ webapp/src/components/AssetPage/BaseDetail/BaseDetail.css
@@ -1,18 +1,3 @@
-.BaseDetail .top-header {
-  display: flex;
-  align-items: center;
-}
-
-.BaseDetail .top-header .back {
-  display: unset;
-  align-self: flex-start;
-}
-
-.BaseDetail .top-header .favorites {
-  flex-grow: 1;
-  justify-content: flex-end;
-}
-
 .BaseDetail .ui.container > .info {
   width: 100%;
   display: flex;
