+++ webapp/src/components/AssetPage/BaseDetail/BaseDetail.types.ts
@@ -14,9 +14,4 @@ export type Props = {
   className?: string
   actions?: ReactNode
   showDetails?: boolean
-  isFavoritesEnabled: boolean
-  onBack: (location?: string) => void
 }
-
-export type MapStateProps = Pick<Props, 'isFavoritesEnabled'>
-export type MapDispatchProps = Pick<Props, 'onBack'>
