+++ webapp/src/components/AssetPage/ItemDetail/ItemDetail.module.css
@@ -1,7 +1,6 @@
 .ItemDetail {
   display: flex;
   flex-direction: column;
-  margin-top: 30px;
 }
 
 .ItemDetail .assetImageContainer :global(.AssetImage) {
