+++ webapp/src/components/AssetPage/OnBack/OnBack.container.ts
@@ -3,8 +3,8 @@ import { Dispatch } from 'redux'
 import { getIsFavoritesEnabled } from '../../../modules/features/selectors'
 import { RootState } from '../../../modules/reducer'
 import { goBack } from '../../../modules/routing/actions'
-import { MapStateProps, MapDispatchProps } from './BaseDetail.types'
-import BaseDetail from './BaseDetail'
+import { MapStateProps, MapDispatchProps } from './OnBack.types'
+import OnBack from './OnBack'
 
 const mapState = (state: RootState): MapStateProps => ({
   isFavoritesEnabled: getIsFavoritesEnabled(state)
@@ -14,4 +14,4 @@ const mapDispatch = (dispatch: Dispatch): MapDispatchProps => ({
   onBack: (location?: string) => dispatch(goBack(location))
 })
 
-export default connect(mapState, mapDispatch)(BaseDetail)
+export default connect(mapState, mapDispatch)(OnBack)
