+++ webapp/src/components/AssetPage/OnBack/index.tsx
@@ -1,0 +1,3 @@
+import OnBack from './OnBack.container'
+
+export default OnBack
